renaming widgets to comply with the rules of page builder
