+++ src/components/Widgets/Block/Block.vue
@@ -119,9 +119,9 @@
 </template>
 
 <script>
-import ProductItem from 'src/components/Widgets/ProductItem/productItem'
+import ProductItem from 'src/components/Widgets/ProductItem/ProductItem'
 import Slider from 'components/Widgets/Slider/Slider'
-import SetItem from 'components/Widgets/SetItem/setItem'
+import SetItem from 'components/Widgets/SetItem/SetItem'
 import ContentItem from 'components/Widgets/ContentItem/ContentItem'
 import { Block } from 'src/models/Block'
 import { dragscrollNext as dragscroll } from 'vue-dragscroll'
