+++ src/components/Widgets/ProductItem/ProductItem.vue
