+++ src/components/Widgets/SetItem/SetItem.vue
